reorder functions in utils file
