+++ pipelines/datasets/br_sfb_sicar/constants.py
@@ -11,8 +11,9 @@ class Constants(Enum): # pylint: disable=c0103
     OUTPUT_PATH = '/tmp/car/output'
 
     UF_SIGLAS = [
-        'AC', 'AL', 'AP', #'AM', 'BA', 'CE', 'DF', 'ES', 'GO', 'MA',
+        'SP',
+        #'AL', 'AP', 'AM', 'BA', 'CE', 'DF', 'ES', 'GO', 'MA',
         #'MT', 'MS', 'MG',
-        'PA', 'PB', 'PR', 'PE', 'PI', 'RJ', 'RN',
-        #'RS', 'RO', 'RR', 'SC', 'SP', 'SE', 'TO'
+        #'PA', 'PB', 'PR', 'PE', 'PI', 'RJ', 'RN',
+        #'RS', 'RO', 'RR', 'SC', 'AC', 'SE', 'TO'
     ]
