+++ pipelines/datasets/br_sfb_sicar/tasks.py
@@ -17,34 +17,12 @@ from pipelines.datasets.br_sfb_sicar.constants import (
 )
 
 from pipelines.datasets.br_sfb_sicar.utils import (
-    process_all_files
+    process_all_files,
+    retry_download_car,
 )
 
 from pipelines.utils.utils import log
 from pipelines.constants import constants
-
-
-#wrapper para usar while e gerenciar erors
-def retry_download_car(car, state, polygon, folder, max_retries=8):
-    retries = 0
-    success = False
-
-    while retries < max_retries and not success:
-        try:
-            car.download_state(state=state, polygon=polygon, folder=folder)
-            success = True
-            log(f'Download do estado {state} concluído com sucesso.')
-        except httpx.ReadTimeout as e:
-            retries += 1
-            log(f'Erro de timeout ao baixar {state}. Tentativa {retries} de {max_retries}. Exceção: {e}')
-            log(f'Tentando novamente em 8 segundos')
-            tm.sleep(8)
-            if retries >= max_retries:
-                log(f'Falha ao baixar o estado {state} após {max_retries} tentativas.')
-                raise e
-        except Exception as e:
-            log(f'Erro inesperado ao baixar {state}: {e}')
-            raise e
 
 
 @task(
